Use uint64 for system call return type
This change is copied from xv6-riscv. It simplifies the handling of
syscall return values which are blindly copied into return register (rax).
This change is on the kernel side only: in user mode, syscall return types
stay the same (int, char * or whatever), the value in rax is taken as is.